requirements.txt
annotated-types==0.6.0
anyio==4.3.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
diskcache==5.6.3
distro==1.9.0
dnspython==2.6.1
docker==7.0.0
eventlet==0.35.2
exceptiongroup==1.2.0
filelock==3.13.1
FLAML==2.1.1
fsspec==2024.2.0
greenlet==3.0.3
gurobipy==11.0.0
h11==0.14.0
httpcore==1.0.4
httpx==0.27.0
huggingface-hub==0.21.1
idna==3.6
Jinja2==3.1.3
MarkupSafe==2.1.5
mpmath==1.3.0
networkx==3.2.1
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.19.3
nvidia-nvjitlink-cu12==12.3.101
nvidia-nvtx-cu12==12.1.105
openai==1.12.0
packaging==23.2
pillow==10.2.0
pyautogen==0.2.15
pycparser==2.21
pydantic==2.6.2
pydantic_core==2.16.3
python-dotenv==1.0.1
PyYAML==6.0.1
regex==2023.12.25
requests==2.31.0
safetensors==0.4.2
sniffio==1.3.1
soundfile==0.12.1
sympy==1.12
termcolor==2.4.0
tiktoken==0.6.0
tokenizers==0.15.2
torch==2.2.1
torchaudio==2.2.1
torchvision==0.17.1
tqdm==4.66.2
transformers==4.38.1
triton==2.2.0
typing_extensions==4.10.0
urllib3==2.2.1
